requirements.txt
# Chihuahua vs Chikuwa Classifier Dependencies

# Core Python packages
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
matplotlib>=3.5.0

# Image search and processing
duckduckgo-search>=3.8.0
opencv-python>=4.5.0

# PyTorch and machine learning
torch>=1.12.0
torchvision>=0.13.0

# Web interface (optional)
gradio>=3.0.0

# Additional utilities (if needed)
pyperclip
youtube_transcript_api
yt-dlp
google-generativeai
